client/requirements.txt
googleapis-common-protos==1.62.0
grpc-interceptor==0.15.4
grpcio==1.60.0
grpcio-status==1.60.0
grpcio-tools==1.59.3
scikit-learn==1.3.2
rich==13.7.0
python-dotenv==0.19.0